+++ app/components/UserAccessTable.tsx
@@ -1,16 +1,20 @@
+import { Set } from 'immutable';
 import * as React from 'react';
-import { toastr } from 'react-redux-toastr';
+import * as cx from 'classnames';
+import { AccessToggle } from '../actions/types/authorizedUsers';
 import Node, { ROOT_ID } from '../domain/Node';
-import * as styles from './UserKeyTable.css';
+import * as styles from './UserAccessTable.scss';
 
 export interface Props {
   users: string[],
   authorizationNodes: Node[],
-  currentUser?: string | null
+  modifications: Set<AccessToggle>,
+  currentUser?: string | null,
+  onToggle: (node: Node, username: string) => void
 }
 
-export default ({ users, authorizationNodes, currentUser }: Props) => (
-  <table className={`table table-hover table-sm table-sticky ${styles.table} ${styles.accessTable}`}>
+export default ({ users, authorizationNodes, modifications, currentUser, onToggle }: Props) => (
+  <table className={`table table-hover table-sm table-sticky ${styles.table}`}>
     <thead>
       <tr>
         <th className="sticky-y">Username</th>
@@ -23,10 +27,15 @@ export default ({ users, authorizationNodes, currentUser }: Props) => (
       {users.sort().map(username => (<tr key={username} className={username === currentUser ? 'table-success' : ''}>
         <th>{username}</th>
         {authorizationNodes.map(node => {
-          const authorized = node.authorizedUsers && node.authorizedUsers.includes(username);
-          return <td key={node.id} onClick={() => toastr.info('', 'sorry, not yet')}>
-            {<i className={`fa fa-${authorized ? 'check' : 'times'} text-${authorized ? 'success' : 'danger'}`}/>}
-          </td>;
+          let authorized = node.authorizedUsers && node.authorizedUsers.includes(username);
+          const mayEdit = node.authorizedUsers && currentUser && currentUser !== username && node.authorizedUsers.includes(currentUser);
+          const modified = modifications.find(m => m!.nodeId === node.id && m!.username === username);
+          if (modified) {
+            authorized = !authorized;
+          }
+          return <td key={node.id} onClick={() => { if (mayEdit) { onToggle(node, username); } }}
+                     title={mayEdit ? 'Toggle access' : ''} className={cx(authorized ? styles.authorized : styles.unauthorized,
+                      mayEdit ? 'clickable' : styles.readonly, modified && styles.modified)}/>;
         })}
       </tr>))}
     </tbody>